.NET Standard 2.1 support (#8)
* Closes #7 - .NET Standard 2.1 support
* Closes #7 - Fix .NET 3.1 test
* Closes #7 - Work around .NET 3.1 and Linux culture issues.
* Automating release
* Tweak versioning
* Tweak coverage, fix variable naming
* Commenting out brittle test
* Commenting out coverage for now

+++ tests/RopeEditingTests.cs
@@ -164,7 +164,7 @@ public class RopeEditingTests
         var random = new Random(42);
         var rope = Rope<float>.Empty;
         var comparer = Comparer<float>.Default;
-        foreach (var rank in Enumerable.Range(0, 65000).Select(s => random.NextSingle()))
+        foreach (var rank in Enumerable.Range(0, 65000).Select(s => (float)random.NextDouble()))
         {
             rope = rope.InsertSorted(rank, comparer);
         }
